+++ start/src/App.vue
@@ -8,6 +8,8 @@
   <f-alert bold close type="warning">警告提示信息</f-alert>
   <f-alert bold close type="info">信息提示信息</f-alert>
 
+  <f-aside background="red">Aside</f-aside>
+
   <!-- <f-alert simple type="default">默认提示信息</f-alert>
   <f-alert simple type="primary">主要提示信息</f-alert>
   <f-alert simple type="success">成功提示信息</f-alert>